lamassat l akhira! cpp09

+++ cpp09/ex01/RPN.hpp
@@ -7,7 +7,7 @@
 class RPN
 {
 private:
-    std::stack<int> _numbers;
+    std::stack<double> _numbers;
 public:
     RPN();
     RPN(const RPN& copy);
